Optimized scripts, eagle work in progress

+++ Assets/Scripts/UI/ScoreManager.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class ScoreManager : MonoBehaviour
 {
     public static ScoreManager instance;
-    private int _score = 0;
+    private ushort _score = 0;
     public TMP_Text scoreText; 
 
     private void Awake()
@@ -15,7 +15,7 @@ public class ScoreManager : MonoBehaviour
         scoreText.text = _score.ToString();
     }
 
-    public void Increase(int points)
+    public void Increase(ushort points)
     {
         _score += points;
         scoreText.text = _score.ToString();
